fix: define the CSRF cookie without the __Host prefix
From https://developer.mozilla.org/en-US/docs/Web/HTTP/Guides/Cookies#cookie_prefixes:

Cookies with names starting with __Host- must be set with the Secure attribute
by a secure page (HTTPS). In addition, they must not have a Domain attribute
specified, and the Path attribute must be set to /. This guarantees that such
cookies are only sent to the host that set them, and not to any other host on
the domain. It also guarantees that they are set host-wide and cannot be
overridden on any path on that host. This combination yields a cookie that is
as close as can be to treating the origin as a security boundary.

This means, that since we want to leverage the Domain attribute, we can't use
the __Host- prefix anymore.